+++ lib/initialization/initializers/kakao_sdk_initializer.dart
@@ -3,6 +3,7 @@ import 'package:kakao_flutter_sdk/kakao_flutter_sdk.dart';
 
 class KakaoSdkInitializer {
   Future<void> initialize() async {
+    print("[KakaoSdkInitializer] initialize");
     await dotenv.load(fileName: "assets/config/keys.env");
     final String nativeAppKey = dotenv.env['KAKAO_SDK_NATIVE_APP_KEY'] ?? 'KAKAO_SDK_NATIVE_APP_KEY';
     final String jsAppKey = dotenv.env['KAKAO_SDK_JS_APP_KEY'] ?? 'KAKAO_SDK_JS_APP_KEY';
