+++ lib/routers/routers.dart
@@ -20,69 +20,82 @@ class Routers {
   static const INITIAL = '/root';
 
   static final routers = [
-    GetPage(name: '/root', page: () => RootScreen(), transition: Transition.noTransition),
     GetPage(
-        name: '/home',
-        page: () => const HomeScreen(),
-        binding: BindingsBuilder(() {
-          Get.lazyPut<HomeCtr>(() => HomeCtr());
-        }),
-        transition: Transition.noTransition),
+      name: '/root',
+      page: () => RootScreen(),
+      transition: Transition.noTransition,
+    ),
     GetPage(
-        name: '/scriptInput/input',
-        page: () => const ScriptInputScreen(),
-        binding: BindingsBuilder(() {
-          Get.lazyPut<ScriptInputCtr>(() => ScriptInputCtr());
-        }),
-        transition: Transition.noTransition),
+      name: '/home',
+      page: () => const HomeScreen(),
+      binding: BindingsBuilder(() {
+        Get.lazyPut<HomeCtr>(() => HomeCtr());
+      }),
+      transition: Transition.noTransition,
+    ),
     GetPage(
-        name: '/scriptInput/result',
-        page: () => const ScriptExpectedTimeScreen(),
-        binding: BindingsBuilder(() {
-          Get.lazyPut<ScriptInputCtr>(() => ScriptInputCtr());
-        }),
-        transition: Transition.noTransition),
+      name: '/scriptInput/input',
+      page: () => const ScriptInputScreen(),
+      binding: BindingsBuilder(() {
+        Get.lazyPut<ScriptInputCtr>(() => ScriptInputCtr());
+      }),
+      transition: Transition.noTransition,
+    ),
     GetPage(
-        name: '/voiceRecodeNoScript',
-        page: () => const VoiceRecodeScreenNoScript(),
-        binding: BindingsBuilder(() {
-          Get.lazyPut<VoiceRecodeCtr>(() => VoiceRecodeCtr());
-        }),
-        transition: Transition.noTransition),
+      name: '/scriptInput/result',
+      page: () => const ScriptExpectedTimeScreen(),
+      binding: BindingsBuilder(() {
+        Get.lazyPut<ScriptInputCtr>(() => ScriptInputCtr());
+      }),
+      transition: Transition.noTransition,
+    ),
     GetPage(
-        name: '/voiceRecodeWithScript',
-        page: () => const VoiceRecodeScreenWithScript(),
-        binding: BindingsBuilder(() {
-          Get.lazyPut<VoiceRecodeCtr>(() => VoiceRecodeCtr());
-        }),
-        transition: Transition.noTransition),
+      name: '/voiceRecodeNoScript',
+      page: () => const VoiceRecodeScreenNoScript(),
+      binding: BindingsBuilder(() {
+        Get.lazyPut<VoiceRecodeCtr>(() => VoiceRecodeCtr());
+      }),
+      transition: Transition.noTransition,
+    ),
     GetPage(
-        name: '/themeInput',
-        page: () => const ThemeInputScreen(),
-        binding: BindingsBuilder(() {
-          Get.lazyPut<ThemeInputCtr>(() => ThemeInputCtr());
-        }),
-        transition: Transition.noTransition),
+      name: '/voiceRecodeWithScript',
+      page: () => const VoiceRecodeScreenWithScript(),
+      binding: BindingsBuilder(() {
+        Get.lazyPut<VoiceRecodeCtr>(() => VoiceRecodeCtr());
+      }),
+      transition: Transition.noTransition,
+    ),
     GetPage(
-        name: '/historyThemeList',
-        page: () => const HistoryScreen(),
-        binding: BindingsBuilder(() {
-          Get.lazyPut<HistoryCtr>(() => HistoryCtr());
-        }),
-        transition: Transition.noTransition),
+      name: '/themeInput',
+      page: () => const ThemeInputScreen(),
+      binding: BindingsBuilder(() {
+        Get.lazyPut<ThemeInputCtr>(() => ThemeInputCtr());
+      }),
+      transition: Transition.noTransition,
+    ),
     GetPage(
-        name: '/historyMajorDetail',
-        page: () => const HistoryMajorDetailScreen(),
-        binding: BindingsBuilder(() {
-          Get.lazyPut<HistoryCtr>(() => HistoryCtr());
-        }),
-        transition: Transition.noTransition),
+      name: '/historyThemeList',
+      page: () => const HistoryScreen(),
+      binding: BindingsBuilder(() {
+        Get.lazyPut<HistoryCtr>(() => HistoryCtr());
+      }),
+      transition: Transition.noTransition,
+    ),
     GetPage(
-        name: '/practiceResult',
-        page: () => const PracticeResultScreen(),
-        binding: BindingsBuilder(() {
-          Get.lazyPut<PracticeResultCtr>(() => PracticeResultCtr());
-        }),
-        transition: Transition.noTransition),
+      name: '/historyMajorDetail',
+      page: () => const HistoryMajorDetailScreen(),
+      binding: BindingsBuilder(() {
+        Get.lazyPut<HistoryCtr>(() => HistoryCtr());
+      }),
+      transition: Transition.noTransition,
+    ),
+    GetPage(
+      name: '/practiceResult',
+      page: () => const PracticeResultScreen(),
+      binding: BindingsBuilder(() {
+        Get.lazyPut<PracticeResultCtr>(() => PracticeResultCtr());
+      }),
+      transition: Transition.noTransition,
+    ),
   ];
 }
